+++ src/broadcasting.jl
@@ -86,7 +86,7 @@ bcastlazy(A, dims::Integer...) =
 
 yields the size `siz` of the array that would result from applying broadcasting
 rules (see `broadcast`) to arguments `A`, `B`, etc.  The result is a tuple of
-integers (of type `Int`).  Call [`check_dimensions`](@ref) if you want to also
+integers (of type `Int`).  Call [`check_size`](@ref) if you want to also
 make sure that the result is a list of valid dimensions.
 
 The method can also be applied to a single dimension:
@@ -98,7 +98,7 @@ throwing an exception if dimensions are not compatible according to
 broadcasting rules.  This is the same as `Base.Broadcasting._bcs1` but it takes
 care of converting to `Int`.
 
-See also [`standard_size`](@ref), [`check_dimensions`](@ref), [`bcastcopy`](@ref),
+See also [`standard_size`](@ref), [`check_size`](@ref), [`bcastcopy`](@ref),
 [`bcastlazy`](@ref).
 
 """