+++ src/utils.jl
@@ -31,8 +31,8 @@ converts `siz` to an `N`-tuple of `Int`.  Argument `siz` can be a scalar
 integer or a tuple of integers.  Argument `siz` is returned if already of the
 correct type.
 
-The union `ArraySize` matches the types of argument `arg` acceptable for
-`to_size(arg)`: scalar integers and tuples of integers.
+The union [`ArraySize`](@ref) matches the types of argument `arg` acceptable
+for `to_size(arg)`: scalar integers and tuples of integers.
 
 This method is intended for fast conversion, call `check_size(siz)` to verify
 that all dimensions in `siz` are nonnegative.
@@ -47,10 +47,6 @@ with `dim1`, `dim2`, ... scalar integers also yields an `N`-tuple of `Int`.
 to_size(siz::Tuple{Vararg{Integer}}) = to_int(siz)
 to_size(siz::Integer) = (to_int(siz),)
 to_size(siz::Integer...) = to_size(siz)
-
-const ArraySize = Union{Integer,Tuple{Vararg{Integer}}}
-
-@doc @doc(to_szie) ArraySize
 
 """