+++ NEWS.md
@@ -1,6 +1,11 @@
-- `dimensions(A::AbstractArray)` has been deprecated in favor of `standard_size(A)` while
-  `dimensions` has been deprecated in favor of `to_size` for other types of arguments.
-  Union `Dimensions` has been deprecated favor of `ArraySize`.
+- `check_dimensions` deprecated in favor of `check_size` which yields the
+  number of elements and throws an `ArgumentError` exception if not all
+  dimensions are nonnegative.
+
+- `dimensions(A::AbstractArray)` has been deprecated in favor of
+  `standard_size(A)` while `dimensions` has been deprecated in favor of
+  `to_size` for other types of arguments.  Union `Dimensions` has been
+  deprecated favor of `ArraySize`.
 
 - New method `to_int` for quick conversion of values to `Int`.